requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
django-environ==0.4.5
djangorestframework==3.12.4
idna==2.10
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
